requirements.txt
django==1.11
djangorestframework
django-celery==3.1.17
celery==3.1.23
gunicorn==19.6.0
psycopg2==2.6.2
requests==2.11.1
